Refactor settings layout and planner display
